Return created and updated songs and players from API functions
The hymnal server is (tentatively) updating to use automatic create and
update timestamps. As a result, after creating or updating a record, the
client needs to know the values assigned by the server in order to
fully populate its local cache. This modifies the `create` and `update`
methods of the API wrappers to return the created or updated object as
stored by the server.